Documents/EmpServe-M-E-System-main/EmpServe-M-E-System-main/EmpServe-M-E-System-main/backend/requirements.txt
asgiref==3.4.1
autopep8==1.6.0
Django==4.0.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.9.0
pytz==2024.2
tzdata==2024.2
sqlparse==0.5.1
pycodestyle==2.12.1
toml==0.10.2



